# perf(generate): increase performance by reducing unnecessary reads (#57)