--- a/src/helpers/checkProperty.test.ts
+++ b/src/helpers/checkProperty.test.ts
@@ -1,28 +1,28 @@
 import { describe, expect, it } from "vitest";
+import { packageType } from "../types/interface";
 import checkProperty from "./checkProperty";
 
 describe("Check Property", () => {
-    const normalPath = "./package.json";
-    const wrongPath = "./mocks/mockPackage.json";
-    const mockPath = "./src/mocks/mock.package.json";
+    const mockPkg: packageType = {
+        dependencies: {},
+        devDependencies: {
+            typescript: "^4.9.4",
+            sass: "^1.58.3",
+        },
+    };
 
     it("should return true if it's a typescript project", async () => {
-        const isTypeScript = await checkProperty("typescript", mockPath);
+        const isTypeScript = await checkProperty("typescript", mockPkg);
         expect(isTypeScript).toBe(true);
     });
 
     it("should return true if it's a scss project", async () => {
-        const isScss = await checkProperty("sass", mockPath);
+        const isScss = await checkProperty("sass", mockPkg);
         expect(isScss).toBe(true);
     });
 
     it("should return false if it's not property project", async () => {
-        const isScss = await checkProperty("sass", normalPath);
+        const isScss = await checkProperty("something", mockPkg);
         expect(isScss).toBe(false);
-    });
-
-    it("should return false if the package doesn't exist in the project", async () => {
-        const isTypeScript = await checkProperty("typescript", wrongPath);
-        expect(isTypeScript).toBe(false);
     });
 });