--- a/src/helpers/checkProperty.ts
+++ b/src/helpers/checkProperty.ts
@@ -1,37 +1,24 @@
-import fs from "fs";
-import path from "path";
-import getRootPath from "./getRootPath.js";
+import { packageType } from "../types/interface.js";
 
 /**
  * Checks whether the given property exists or not
  *
  * @param property property to check
- * @param packagePath path to package.json
+ * @param pkg package.json contents
  * @returns true if exists or false
  */
 const checkProperty = async (
     property: string,
-    packagePath: string,
+    pkg: packageType,
 ): Promise<boolean> => {
-    const rootPath = getRootPath(process.cwd());
-    const packageJsonPath = path.resolve(rootPath, packagePath);
-
-    try {
-        const data = await fs.promises.readFile(packageJsonPath, "utf-8");
-        const packageJson = JSON.parse(data);
-        const dependencies = packageJson.dependencies || {};
-        const devDependencies = packageJson.devDependencies || {};
-
-        if (
-            Object.hasOwn(dependencies, property) ||
-            Object.hasOwn(devDependencies, property)
-        ) {
-            return true;
-        }
-        return false;
-    } catch (error) {
-        return false;
+    const { dependencies, devDependencies } = pkg;
+    if (
+        Object.hasOwn(dependencies, property) ||
+        Object.hasOwn(devDependencies, property)
+    ) {
+        return true;
     }
+    return false;
 };
 
 export default checkProperty;