--- a/src/helpers/index.ts
+++ b/src/helpers/index.ts
@@ -1,5 +1,7 @@
 import checkReact from "./checkReact.js";
+import getConfig from "./getConfig.js";
 import getRootPath from "./getRootPath.js";
+import getPackageFile from "./getPackageFile.js";
 import checkProperty from "./checkProperty.js";
 
-export { checkReact, checkProperty, getRootPath };
+export { checkReact, checkProperty, getConfig, getRootPath, getPackageFile };