<!-- Thank you for contributing! -->

### Description

- removed unnecessary file reads on `package.json`
- added method `getPackgeFile` to read `package.json` once and return it's content
- updated `checkReact` and `checkProperty` to use the content retrieved from `getPackageFile`
- updated the tests for `checkReact` and `checkProperty`
- added  test for `getPackageFile`.

### Additional context

Before these changes it took an average of `2.23ms  - 2.7ms` to generate basic component, after these changes reduced the time to an average of `1.2ms - 1.7ms`.

### What is the purpose of this pull request? <!-- (put an "X" next to an item) -->

-   [ ] Bug fix
-   [ ] New Feature
-   [X] Other

### Before submitting the PR, please make sure you do the following

-   [X] Read the [Contributing Guidelines](https://github.com/arclix/core/blob/master/CONTRIBUTING.md).
-   [X] Read the [Pull Request Guidelines](https://github.com/arclix/core/blob/master/CONTRIBUTING.md#pull-request-guidelines) and follow the [PR Title Convention](https://github.com/arclix/core/blob/master/.github/COMMIT_CONVENTION.md).
-   [X] Check that there isn't already a PR that solves the problem the same way to avoid creating a duplicate.
-   [X] Provide a description in this PR that addresses **what** the PR is solving, or reference the issue that it solves (e.g. `fixes #123`).
-   [X] Ideally, include relevant tests that fail without this PR but pass with it.